requirements.txt
django==5.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
pytz==2024.1
python-dotenv==1.0.1
aio-georss-gdacs